Enhance Input Area Shipping

+++ src/Com.Danliris.Service.Packing.Inventory.Application/ToBeRefactored/DyeingPrintingAreaInput/Shipping/InputShippingProductionOrderViewModel.cs
@@ -21,7 +21,7 @@ namespace Com.Danliris.Service.Packing.Inventory.Application.ToBeRefactored.Dyei
         public bool HasOutputDocument { get; set; }
         public string Grade { get; set; }
         public string Packing { get; set; }
-        public double QtyPacking { get; set; }
+        public decimal QtyPacking { get; set; }
         public double Qty { get; set; }
         public string PackingType { get; set; }
     }
